UADS-18 refactor: abstract event card component
