+++ web/src/pages/Home.tsx
@@ -14,6 +14,7 @@ import { useState, useEffect } from "react";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Slider from "react-slick";
+import EventCard from "@components/EventCard";
 
 //TODO: Get actual info from db or cms
 const sponsorsDummyData: string[] = [
@@ -223,6 +224,7 @@ function SignupSection() {
   );
 }
 
+//TODO: Get actual info from db or cms
 function EventSlider() {
   const settings = {
     dots: true,
@@ -295,47 +297,6 @@ function EventSection() {
       </h1>
       <EventSlider />
       <PinkButton linkto="/events" buttontext="See More Events" />
-    </div>
-  );
-}
-
-interface EventCardProps {
-  image: string;
-  eventname: string;
-  eventdesc: string;
-  month: string;
-  date: string;
-}
-
-function EventCard({
-  image,
-  eventname,
-  eventdesc,
-  month,
-  date,
-}: EventCardProps) {
-  console.log(image);
-  return (
-    <div className=" mx-10 flex flex-col h-80 bg-white rounded-3xl overflow-hidden relative z-10">
-      <img
-        src={image}
-        className="flex object-cover h-full w-full absolute z-0 blur-sm"
-      />
-      <div className="flex overflow-hidden h-56 relative justify-center ">
-        <img src={image} className="flex object-contain z-10" />
-      </div>
-      <div className="flex flex-row w-full z-10 grow">
-        <div className="flex flex-col w-1/6 p-4 bg-white justify-center items-center text-black">
-          <h1>{month}</h1>
-          <h1 className="font-bold">{date}</h1>
-        </div>
-        <div className="flex flex-col p-4 bg-pink w-full text-light-pink">
-          <h1 className="font-bold">{eventname}</h1>
-          <h1 className="text-xs xl:text-base overflow-hidden text-ellipsis">
-            {eventdesc}
-          </h1>
-        </div>
-      </div>
     </div>
   );
 }
